requirements.txt
# requirements.txt
django
djangorestframework
djangorestframework-simplejwt